+++ tests/test_package/inspect_package/approvers/renamer.py
@@ -1,0 +1,20 @@
+from copy import copy
+
+from inspect_ai.approval import Approval, Approver, approver
+from inspect_ai.solver import TaskState
+from inspect_ai.tool import ToolCall, ToolCallView
+
+
+@approver
+def renamer(function_name: str) -> Approver:
+    async def approve(
+        message: str,
+        call: ToolCall,
+        view: ToolCallView,
+        state: TaskState | None = None,
+    ) -> Approval:
+        call = copy(call)
+        call.function = function_name
+        return Approval(decision="modify", modified=call)
+
+    return approve
